fix(profile): addressed comments on branch and brought back download function, linted code
